Adding test to ensure partition reassignment don't like on maintenance brokers
